Add env var to skip shrink step

+++ centos-7.sh
@@ -31,11 +31,13 @@ find /var/log -type f -exec truncate --size=0 {} \;
 truncate --size=0 /etc/machine-id
 
 # Shrink disk
-dd if=/dev/zero of=/boot/ZERO bs=1M || echo "ignoring expected dd error"
-rm /boot/ZERO
-dd if=/dev/zero of=/ZERO bs=1M  || echo "ignoring expected dd error"
-rm /ZERO
-sync
+if [[ -z "$PACKER_SKIP_SHRINK_DISK" ]]; then
+  dd if=/dev/zero of=/boot/ZERO bs=1M || echo "ignoring expected dd error"
+  rm /boot/ZERO
+  dd if=/dev/zero of=/ZERO bs=1M  || echo "ignoring expected dd error"
+  rm /ZERO
+  sync
+fi
 
 export HISTSIZE=0
 rm -f /root/.wget-hsts
